refix damage, add config
+ added config to enable optional legacy difficulty settings
+ hospital zombie now have a 25% chance to drop healing potions
~ increased the droprate of rare loot from firemen 5% +> 10%
~ increased fireman hp 20 -> 30
~ fixed issue that caused firemen to have the wrong armour value
~ decreased fireman armour from the fixed value 20 -> 10
~ changed gas detection to be based off of eye position instead of feet position
~ fixed bug that caused zombies to ignore weapons and armour when dealing damage
~ fixed bug that caused entities other than normal zombies to not be slowed during the day
~ fixed bug that prevented ambushes from spawning
~ rewrote zombie fireman renderer to be less glitchy
- removed the ability to place hot liquids
- removed the ability for dispensers to place lava buckets
- lava no longer melts through buckets

+++ src/main/java/net/smileycorp/ldoh/client/entity/model/ModelTurret.java
@@ -142,7 +142,7 @@ public class ModelTurret extends ModelBase {
 		if (entity != null) {
 			if (entity.getTeam()!=null) {
 				int colour = Minecraft.getMinecraft().fontRenderer.getColorCode(entity.getTeam().getColor().formattingCode);
-				GlStateManager.color((float)(colour >> 16) / 255.0F, (float)(colour >> 8 & 255) / 255.0F, (float)(colour & 255) / 255.0F);
+				GlStateManager.color((colour >> 16) / 255.0F, (colour >> 8 & 255) / 255.0F, (colour & 255) / 255.0F);
 			} else {
 				GlStateManager.color(0.25f, 0.25f, 0.25f);
 			}
